refactor: 수식 표현 방식 수정

+++ backend/src/main/java/com/mapbefine/mapbefine/common/filter/LatencyRecorder.java
@@ -17,7 +17,7 @@ public class LatencyRecorder {
 
         threadLocal.remove();
 
-        return (double) (end - start) / 1000;
+        return (end - start) / 1000d;
     }
 
 }
